Decompose cost scanner tests
